pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"

requires = ["poetry-core=1.1.3"]


[tool.poetry]
name = "snowflake-extentions"

description = ""

version = "0.1.0"

authors = []


[tool.poetry.dependencies]
# https://github.com/python-poetry/poetry/issues/2444#issuecomment-632471356
python = ">=3.11, <3.12"
snowflake-snowpark-python = "^1.18.0"
